fixing reload after searching new city
