+++ src/components/WeatherForm.module.css
@@ -8,7 +8,7 @@
 
 .card-alignment {
   margin: 0 auto;
-  margin-top: 10vh;
+  margin-top: 5vh;
   padding: 1.2rem;
   width: 40rem;
   min-width: 5rem;
